refactor: 비디오템플릿 일반화 및 오타 수정
